Refactor: pull method up to parent
